api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
supabase>=2.0.0
